+++ src/rules/required.js
@@ -4,13 +4,15 @@
  * Checks that a value is present for the field being validated
  *
  */
+import $ from '@justeat/f-dom';
+
 export default {
     condition: field => field.hasAttribute('required') || field.hasAttribute('data-val-required'),
 
     test: field => {
         // Required checkbox & radio, 1 should be checked.
         if (field.type === 'radio') {
-            const radioInputs = document.querySelectorAll(`[name='${field.name}']:checked`);
+            const radioInputs = $(`[name='${field.name}']:checked`);
             return radioInputs.length > 0;
         }